feat(wallet-login): add `addWalletLogin` and `removeWalletLogin` w
