+++ src/mutations/user/walletLogin.ts
@@ -5,12 +5,8 @@ import type {
 } from 'definitions'
 
 import { invalidateFQC } from '@matters/apollo-response-cache'
-import { recoverPersonalSignature } from 'eth-sig-util'
-import { Contract, utils } from 'ethers'
-import { Knex } from 'knex'
 
 import {
-  BLOCKCHAIN_CHAINID,
   NODE_TYPES,
   VERIFICATION_CODE_STATUS,
   VERIFICATION_CODE_TYPE,
@@ -20,20 +16,16 @@ import {
   CodeExpiredError,
   CodeInactiveError,
   CodeInvalidError,
-  CryptoWalletExistsError,
   EmailExistsError,
   EthAddressNotFoundError,
   UserInputError,
 } from 'common/errors'
-import {
-  getAlchemyProvider,
-  getViewerFromUser,
-  IERC1271,
-  setCookie,
-} from 'common/utils'
+import { getViewerFromUser, setCookie } from 'common/utils'
 import { redis } from 'connectors'
 
-const resolver: GQLMutationResolvers['walletLogin'] = async (
+const sigTable = 'crypto_wallet_signature'
+
+export const walletLogin: GQLMutationResolvers['walletLogin'] = async (
   _,
   { input: { ethAddress, nonce, signedMessage, signature, email, codeId } },
   context
@@ -46,81 +38,19 @@ const resolver: GQLMutationResolvers['walletLogin'] = async (
     knex,
   } = context
 
-  if (!ethAddress || !utils.isAddress(ethAddress)) {
-    throw new UserInputError('address is invalid')
-  }
-  const sig_table = 'crypto_wallet_signature'
-
-  const lastSigning = await atomService.findFirst({
-    table: sig_table,
-    where: (builder: Knex.QueryBuilder) =>
-      builder
-        .where({ address: ethAddress, nonce })
-        .whereNull('signature')
-        .whereRaw('expired_at > CURRENT_TIMESTAMP'),
-    orderBy: [{ column: 'id', order: 'desc' }],
+  const lastSigning = await userService.verifyWalletSignature({
+    ethAddress,
+    nonce,
+    signedMessage,
+    signature,
   })
-
-  if (!lastSigning) {
-    throw new EthAddressNotFoundError(
-      `wallet signing for "${ethAddress}" not found`
-    )
-  }
-
-  // if it's smart contract wallet
-  const isValidSignature = async () => {
-    const MAGICVALUE = '0x1626ba7e'
-
-    const chainType = 'Polygon'
-
-    const chainNetwork = 'PolygonMainnet'
-
-    const provider = getAlchemyProvider(
-      Number(BLOCKCHAIN_CHAINID[chainType][chainNetwork])
-    )
-
-    const bytecode = await provider.getCode(ethAddress.toLowerCase())
-
-    const isSmartContract = bytecode && utils.hexStripZeros(bytecode) !== '0x'
-
-    const hash = utils.hashMessage(signedMessage)
-
-    if (isSmartContract) {
-      // verify the message for a decentralized account (contract wallet)
-      const contractWallet = new Contract(ethAddress, IERC1271, provider)
-      const verification = await contractWallet.isValidSignature(
-        hash,
-        signature
-      )
-
-      const doneVerified = verification === MAGICVALUE
-
-      if (!doneVerified) {
-        throw new UserInputError('signature is not valid')
-      }
-    } else {
-      // verify signature for EOA account
-      const verifiedAddress = recoverPersonalSignature({
-        data: signedMessage,
-        sig: signature,
-      }).toLowerCase()
-
-      if (ethAddress.toLowerCase() !== verifiedAddress) {
-        throw new UserInputError('signature is not valid')
-      }
-    }
-  }
-
-  await isValidSignature()
-
-  const user = await userService.findByEthAddress(ethAddress)
 
   /**
    * Link
    */
   if (viewer.id && viewer.token && !viewer.ethAddress) {
     await atomService.update({
-      table: sig_table,
+      table: 'crypto_wallet_signature',
       where: { id: lastSigning.id },
       data: {
         signature,
@@ -129,21 +59,7 @@ const resolver: GQLMutationResolvers['walletLogin'] = async (
       },
     })
 
-    if (user) {
-      throw new CryptoWalletExistsError('eth address already has a user')
-    }
-
-    await userService.baseUpdate(viewer.id, {
-      updatedAt: knex.fn.now(),
-      ethAddress: ethAddress.toLowerCase(), // save the lower case ones
-    })
-
-    // archive crypto_wallet entry
-    await atomService.update({
-      table: 'crypto_wallet',
-      where: { userId: viewer.id, archived: false },
-      data: { updatedAt: knex.fn.now(), archived: true },
-    })
+    await userService.addWallet(viewer.id, ethAddress)
 
     await invalidateFQC({
       node: { type: NODE_TYPES.User, id: viewer.id },
@@ -175,7 +91,7 @@ const resolver: GQLMutationResolvers['walletLogin'] = async (
 
     // update crypto_wallet_signature record
     await atomService.update({
-      table: sig_table,
+      table: sigTable,
       where: { id: lastSigning.id },
       data: {
         signature,
@@ -187,6 +103,8 @@ const resolver: GQLMutationResolvers['walletLogin'] = async (
 
     return { token, auth: true, type, user }
   }
+
+  const user = await userService.findByEthAddress(ethAddress)
 
   if (user) {
     // login
@@ -255,4 +173,24 @@ const resolver: GQLMutationResolvers['walletLogin'] = async (
   return tryLogin(AUTH_RESULT_TYPE.Signup)
 }
 
-export default resolver
+export const addWalletLogin: GQLMutationResolvers['addWalletLogin'] = async (
+  _,
+  { input: { ethAddress, nonce, signedMessage, signature } },
+  { viewer, dataSources: { userService } }
+) => {
+  if (viewer.ethAddress) {
+    throw new UserInputError('User has already linked a wallet')
+  }
+  await userService.verifyWalletSignature({
+    ethAddress,
+    nonce,
+    signedMessage,
+    signature,
+  })
+  return await userService.addWallet(viewer.id, ethAddress)
+}
+
+export const removeWalletLogin: GQLMutationResolvers['removeWalletLogin'] =
+  async (_, __, { viewer, dataSources: { userService } }) => {
+    return await userService.baseUpdate(viewer.id, { ethAddress: null })
+  }
